Multi-node parallel option
Changed mc.cores to allow for multi-node processing. There are now 3 options for mc.cores: (1) an object of type "cluster", which can be a multi-node cluster; (2) an integer, which will be used as the number of cores for a single-node cluster; (3) NA [or any other class], which will result in a single-node cluster having the number of cores detected by parallel::detectCores. #3 is an implementation of a to-do item.

If a cluster is passed as mc.cores, the cluster isn't stopped after processing takes place.

In addition, this commit implements the to-do item where the topics vector is sorted in descending order for optimal parallelization.