Expose scoring algorithms to public namespace
Added scoring algorithms to the public namespace so that externally-calculated LDA objects can be scored directly -- this is necessary because the FindTopicsNumber function doesn't currently have the capability of saving the LDA models as they are calculated.

Add roxygen documentation necessary to expose scoring algorithms to public namespace; updated documentation per new roxygen version.

Propose to increment version since it's been a while.