Add return_models option
return_models=TRUE checks for tibble package and if present, returns an additional column in the result dataframe with the LDA model. There is a new demo that shows usage. tibble has been added to Suggests in the DESCRIPTION file.

Other bug fixes to address new documentation requirements in order to pass checks. Reverts a change from the previous commit that inadvertently broke the Griffiths2004 calculation.