Error checking, documentation
Added a check to make sure the logLik vector in the LDA model is > 0 in the Griffiths2004 function. If not, an error message is displayed and NaN is returned as the value. Added documentation stating that in order to use the Griffiths2004 scoring algorithm, the logLik verctor must be retained using the keep control parameter.